ARM: OMAP: board-rm680: experiment with upstream Atmel mXT driver
Note: doesn't work

Is able to communicate with the chip but chip doesn't generate interrupts
so most probably something is missing. As the Harmattan driver is much
larger than the upstream version and since the specifications are not
available it is not easy to know what is missing in order to be able to use
the driver on N9.

Signed-off-by: Jarkko Nikula <jarkko.nikula@jollamobile.com>